@@ -1,31 +1,32 @@
 {
   "private": true,
-  "name": "playnode-one",
+  "name": "playnode",
   "version": "0.0.1",
   "dependencies": {
-    "dompurify": "^2.0.7",
+    "dompurify": "^2.2.2",
     "lodash.clone": "^4.5.0",
     "lodash.merge": "^4.6.2",
-    "object.values": "^1.0.4",
-    "pako": "^1.0.6",
-    "prop-types": "^15.6.2",
-    "radium": "^0.19.6",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "react-key-handler": "^0.3.0",
-    "react-router-dom": "^4.3.1",
-    "react-waypoint": "^7.3.4"
+    "object.values": "^1.1.2",
+    "pako": "^2.0.2",
+    "prop-types": "^15.7.2",
+    "radium": "^0.26.1",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-key-handler": "^1.1.0",
+    "react-router-dom": "^5.2.0",
+    "react-waypoint": "^9.0.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
+    "@babel/cli": "^7.11.6",
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.10.4",
+    "babel-loader": "^8.1.0",
     "http-server": "^0.12.3",
-    "npm-run-all": "^4.1.3",
-    "trash-cli": "^1.4.0",
-    "webpack": "^3.12.0"
+    "npm-run-all": "^4.1.5",
+    "trash-cli": "^4.0.0",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
   },
   "scripts": {
     "quick-start": "npm-run-all build --parallel serve build:watch",
@@ -42,8 +43,8 @@
   },
   "babel": {
     "presets": [
-      "es2015",
-      "react"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ]
   }
 }
